+++ src/core/server/extensions/playerFunctions/setter.ts
@@ -75,7 +75,12 @@ async function firstConnect(p: alt.Player): Promise<void> {
         return;
     }
 
-    const pos = { ...DEFAULT_CONFIG.CHARACTER_CREATOR_POS };
+    if (process.env.ATHENA_READY === 'false') {
+        p.kick('Still warming up...');
+        return;
+    }
+
+    const pos = { ...DEFAULT_CONFIG.CHARACTER_SELECT_POS };
 
     p.dimension = p.id + 1; // First ID is 0. We add 1 so everyone gets a unique dimension.
     p.pendingLogin = true;