+++ src/core/server/interface/DatabaseCollections.ts
@@ -1,0 +1,5 @@
+export enum Collections {
+    Accounts = 'accounts',
+    Characters = 'characters',
+    Options = 'options'
+}