+++ src/core/shared/configurations/shared.ts
@@ -9,5 +9,7 @@ export const SHARED_CONFIG = {
     MAX_PICKUP_RANGE: 2,
     // Interaction
     INTERACTION_ALWAYS_ON: false, // Determines whether or not a player needs to hit `alt` to turn on interaction.
-    MAX_INTERACTION_RANGE: 8
+    MAX_INTERACTION_RANGE: 8,
+    // Voice
+    VOICE_ON: false
 };